requirements.txt
streamlit
pydantic
ddgs
google-genai
openai
python-dotenv
